Добавил визуализацию с SFML

+++ lw4/Shapes/draft/PictureDraft.cpp
@@ -13,7 +13,7 @@ void CPictureDraft::AddShape(std::unique_ptr<CShape>&& shape)
 	m_shapes.push_back(move(shape));
 }
 
-CShape& CPictureDraft::GetShape(unsigned index) const
+CShape& CPictureDraft::GetShape(size_t index) const
 {
 	if (index >= m_shapes.size())
 		throw invalid_argument("requested index is out of range");
@@ -21,7 +21,7 @@ CShape& CPictureDraft::GetShape(unsigned index) const
 	return *m_shapes[index].get();
 }
 
-unsigned CPictureDraft::GetShapesCount() const
+size_t CPictureDraft::GetShapesCount() const
 {
 	return m_shapes.size();
 }
